Add modular base AI script and refactor NPC combat scripts
