+++ scripts/guard_patrol.py
@@ -1,25 +1,29 @@
-from random import choice
-from typeclasses.scripts import Script
+from .combat_ai import BaseCombatAI
 
-class GuardPatrol(Script):
+class GuardPatrol(BaseCombatAI):
     """Walk a patrol path and attack wanted players."""
 
     def at_script_creation(self):
+        super().at_script_creation()
         self.key = "guard_patrol"
         self.desc = "Guard patrol behavior"
         self.interval = 10
-        self.persistent = True
         self.db.patrol = []
         self.db.index = 0
 
-    def at_repeat(self):
+    def select_target(self):
+        npc = self.obj
+        if not npc or not npc.location:
+            return None
+        for obj in npc.location.contents:
+            if obj.tags.get("wanted"):
+                return obj
+        return None
+
+    def move(self):
         npc = self.obj
         if not npc or not npc.location:
             return
-        for obj in npc.location.contents:
-            if obj.tags.get("wanted"):
-                npc.execute_cmd(f"kill {obj.key}")
-                return
         path = self.db.patrol
         if path:
             exit_name = path[self.db.index % len(path)]
@@ -27,8 +31,5 @@ class GuardPatrol(Script):
             exit_obj = npc.search(exit_name, location=npc.location)
             if exit_obj:
                 exit_obj.at_traverse(npc, exit_obj.destination)
-        else:
-            exits = npc.location.contents_get(content_type="exit")
-            if exits:
-                exit_obj = choice(exits)
-                exit_obj.at_traverse(npc, exit_obj.destination)
+                return
+        super().move()
